:hammer: [chore] PhoneNumberAuthView 수정 #129
